Remove sections from motivations, add implementation considerations.
I think the motivations section as-is is too long. It spends a lot of time
talking about what are effectively the tradeoffs. IMO the motivations section
should specify why it's useful to have the sender control the receiver's ACK
frequency, and then we can have a separate section for the tradeoffs.

This takes a stab at that section, lifting things from motivations and adding
some new text specifically addressing the nascent stages of a connection.