Explain how to use the Reordering threshold
fixes #140

Previous attempts to describe the use of the Reordering threshold, tried to specify a continuous range where this condition would be true (e.g. if gaps are in the  range [Last_Acked, Last_Received - Reorder_Threshold]. This means that you would keep sending ACKs for each packet as long as the re-ordering/loss event is on-going. I don't think this should be the intention. Instead my understanding would be that you send one additional ack and then go back to your usual ack strategy. This is what this PR tries to explain.

This is also inline with the behaviour in RFC9000:
> when the packet has a packet number larger than the highest-numbered ack-eliciting packet that has been received and there are missing packets between that packet and this packet.

This mean in RFC9000 you send an ACK when a gap is detected but as soon as you have send the ACK, the highest-numbered ack-eliciting packet is after the gap and you don't send any additional ACKs for this gap.

I think any other behaviour would actually be problematic as the sender can send packets with gaps in the packet number space without a loss, which means the gap would also never be closed.